Tweaks dirt to cover the bottom of the screen.
-There used to be a pixel-ish of space between the dirt sprite and the
bottom of the screen.

+++ Unity_Bloom/Assets/World Controller/BackgroundGenerator.cs
@@ -76,7 +76,7 @@ public class BackgroundGenerator : MonoBehaviour {
 		                                    1.0f);
 
 		// Move it.
-		int yPos = (int)(-(Screen.height / 2) + (dirtObject.localScale.y / 2));
+		float yPos = (int)(-(Screen.height / 2) + (dirtObject.localScale.y / 2));
 		dirtObject.position = new Vector3(0.0f, yPos, dirtDepth);
 	}
